Factor out _get_schedule and _get_ship_config
